+++ buession-json/src/main/java/com/buession/json/strategy/QqSensitiveStrategy.java
@@ -19,13 +19,14 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2023 Buession.com Inc.														       |
+ * | Copyright @ 2013-2024 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 package com.buession.json.strategy;
 
+import com.buession.core.utils.StringUtils;
+
 import java.util.function.Function;
-import java.util.regex.Pattern;
 
 /**
  * QQ 号码脱敏策略
@@ -34,8 +35,6 @@ import java.util.regex.Pattern;
  * @since 2.3.1
  */
 public class QqSensitiveStrategy extends AbstractSensitiveStrategy {
-
-	private final static Pattern PATTERN = Pattern.compile("(([1-9]\\d{2})\\d{2,5}(\\d{2}))|(([1-9]\\d)\\d{2,3}(\\d))");
 
 	/**
 	 * 构造函数
@@ -49,7 +48,7 @@ public class QqSensitiveStrategy extends AbstractSensitiveStrategy {
 
 	@Override
 	public Function<String, String> getFunction() {
-		return str->PATTERN.matcher(str).replaceAll("$2$5" + getReplacement() + "$3$6");
+		return (str)->StringUtils.left(str, 2) + getReplacement() + StringUtils.right(str, str.length() <= 5 ? 1 : 2);
 	}
 
 }